From:    Aurelien Jarno <aurelien@aurel32.net> Subject: [usbutils] Patch to allow empty line in usb mapfiles Date:    Tue, 4 May 2004 14:26:44 +0200
Hello Thomas,

As you may know, usbmodules doesn't like empty lines in usb mapfiles.
When the files are automatically generated by the kernel tools, it is
not a problem. However, for handwritten files there is sometimes an
empty line at the end of the file.

That's why I propose to ignore such lines, that's the purpose of the
attached patch (against the current CVS). Feel free to apply it or not.

Bye,
Aurelien