Fix build issue with libusb location
We recently decided to give libusb a bump from 0.1.12 to 1.0.3 (yeah,
that was woefully ignored) but noticed that usbutils-0.84 broke...
mostly because it couldn't find "usb.h" (the libusb version, not the
<linux/usb.h> version, I take it).

We tried to bump to 0.86, but the issue hasn't gone away.

When building, we need to have either either the libusb-0.1.pc
installed, or else point the LIBUSB_1_0_CFLAGS etc.  We went with the
latter.

But when libusb-compat-0.1.3/libusb gets built, it doesn't seem to pass
these in...  So the "libusb" shim inside libusb-compat can't find
"xxx/usr/include/libusb-1.0/usb.h" ...

I suspect it might have to do with the fact that our $CFLAGS includes
"-iprefix=$(STAGING_DIR) -iwithprefix /include -iwithprefix
/usr/include" which is getting patched in *after* LIBUSB_1_0_CFLAGS...

Also, I noticed that usbutils doesn't actually seem to make use of
LIBUSB_CFLAGS anywhere in the makefile, which means that whatever CFLAGS
you pass in also has to contain whatever LIBUSB_CFLAGS contained...
which is counter-intuitive.

This patch fixes our build issues.