Renamed the package from hid-dump to usbhid-dump
In preparation for integration with usbutils, the package is given a more
globally correct name "usbhid-dump". All the relevant directories and files
are renamed. All the relevant symbols in the code are renamed. The resulting
binary is renamed. All the license notices in the files are updated with the
new name.