lsusb: Show USB 3.0 U1, U2, and LTM status.
The Get Status value for USB 3.0 devices adds three new status bits.
Bits 2 and 3 indicate if U1 and U2 are enabled, and bit 4 indicates if
Latency Tolerance Messaging (LTM) is enabled.

Upcoming patches for the Linux kernel will enable LTM, so lsusb should
be changed to show whether LTM is enabled for a device.

Signed-off-by: Sarah Sharp <sarah.a.sharp@linux.intel.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>