lsusb: Fix LTM capabilities reporting.
The original code to print bmAttributes had the meaning of the LTM bit
backwards.  The USB 3.0 spec, section 9.6.2.2, says that bit 1 means:

"LTM Capable. A value of one in this bit location indicates that this
device has is capable of generating Latency Tolerance Messages."

Only print LTM supported if bit 1 is set.

Signed-off-by: Sarah Sharp <sarah.a.sharp@linux.intel.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>