lsusb: Use vendor and product name fallback logic in -D mode too
commit dddb696e0172 ("lsusb: Read unkown names from sysfs device desc.")
introduced a fallback option for vendor and product names in device
summary lines, fetching them from sysfs when the VID/PID are not present
in udev hwdb. However, it applied this new logic only to summary lines
listing multiple devices (i.e. not -D mode, which only lists a single
device). Refactor the logic into a common function and use it in both
places.

Fixes: dddb696e0172 ("lsusb: Read unkown names from sysfs device desc.")
Signed-off-by: Thomas Hebb <tommyhebb@gmail.com>