lsusb: Fix buffer size copy/paste error
When I added these calls, I mistakenly passed them all the size of the
vendor buffer instead of the buffers they actually fill. As it happens,
all the sizes are the same, so there's no correctness issue, but we
should still fix the issue for code clarity and ease of modification.

Signed-off-by: Thomas Hebb <tommyhebb@gmail.com>