names: simplify get_vendor_product_with_fallback() a bit
Based on some review from Lonnie Abelbeck, if get_sysfs_name() failed in
get_vendor_product_with_fallback(), the strings "[unknown]" would not be
properly set for the product/vendor.

Now the odds of that happening are very slim, but simplify this all by
setting the strings first, and then, if we can find a real override, use
that instead.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>