usbutils: convert build system to use meson
Autotools is great for what it is designed for, but that epoch is long
ago, and shipping random files/scripts that are not actually checked
into the build system is ripe for problems.

Convert the build over to use meson instead.  Bonus is that the build
logic is simpler (if not still messy in places), and that the build
itself is way faster.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>