meson: add all* CFLAGS from travis-autogen.sh
Add all the flags, apart from -Werror and -fdiagnostics-color=auto.

We don't want for the former in the default build, but for developer,
maintainer and/or CI builds ... Some patches adding CI, fixing the
warnings, etc will be coming later on.

The latter overrides the default color handling, so we cannot see them
pretty colo(u)rs ;-P

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>